requirements.txt
python==3.10.13 
tokenizers==0.15.0
transformers==4.36.2
tqdm==4.66.1
streamlit==1.29.0
scikit-learn==1.2.2 
regex==2023.12.25
pandas==2.1.4 
matplotlib-base==3.8.2          
matplotlib-inline==0.1.6
lightgbm==4.2.0 
joblib==1.3.2
huggingface_hub==0.20.0
catboost==1.2.2 
datasets==2.16.0
pypdf2==2.11.1